fix stuck process detection
a maxsize on the thread pool's queue allows gauging the worker's
vitality; if the queue is full, we should try to put the last accepted
connection on it until there's room or the parent kills us.  this is
done in a busy loop, so too small a size could incur significant
overhead.

should queue-maxsize be the same length as the socket's backlog?